[Server] Add cast

+++ src/common/map/map_orientation.c
@@ -33,11 +33,11 @@ uint8_t map_rotate_orientation(cardinal_dir_t dir, uint8_t orientation)
 	case (NORTH):
 		return (orientation);
 	case (EAST):
-		return ((orientation + 1) % 8 + 1);
+		return ((uint8_t) ((orientation + 1) % 8 + 1));
 	case (SOUTH):
-		return ((orientation + 3) % 8 + 1);
+		return ((uint8_t) ((orientation + 3) % 8 + 1));
 	case (WEST):
-		return ((orientation + 5) % 8 + 1);
+		return ((uint8_t) ((orientation + 5) % 8 + 1));
 	}
 	return (orientation);
 }
